Input formats

Input formats

Currently supported input file formats with different models:

File type	Surface model	Volume model	Charges included

HMO	✓		✓

Mcsf		✓

MSMS	✓

OFF	✓

PQR			✓ (charges only)

HMO

NESSie.Format.readhmo — Function.

readhmo{T <: AbstractFloat}(
    stream::IOStream,
          ::Type{T}=Float64
)

Reads a complete surface model from the given HMO file.

Return type

Model{T, Triangle{T}}

Alias

readhmo{T}(fname::String, ::Type{T}=Float64)

Reads the model using a file name rather than a IOStream object.

source

Mcsf

NESSie.Format.readmcsf — Function.

readmcsf{T <: AbstractFloat}(
    stream::IOStream,
          ::Type{T}=Float64;
    # kwargs
    domain::Symbol=:none
)

Reads a volume model from the given GAMer-generated mcsf file.

Note

This file type does not support charge models! Hence, the charge list of the returning Model object is empty and has to be set separately.

Arguments

domain Element domain (solute :Ω, solvent :Σ, or :none)

Return type

Model{T, Tetrahedron{T}}

Aliases

readmcsf{T}(
    fname ::String,
          ::Type{T}=Float64;
    # kwargs
    domain::Symbol=:none
)

Reads the model using a file name rather than a IOStream object.

readmcsf{T}(
    fnameΩ::String,
    fnameΣ::String,
          ::Type{T}=Float64
)

Reads the contents of two separate files (given by name), sets the element domains to :Ω or :Σ, respectively, and returns a single (merged) Model object.

source

MSMS

NESSie.Format.readmsms — Function.

readmsms{T <: AbstractFloat}(
    vertstream::IOStream,
    facestream::IOStream,
              ::Type{T}=Float64
)

Reads a surface model from the given MSMS-generated .face and .vert files.

Note

This file type does not support charge models! Hence, the charge list of the returning Model object is empty and has to be set separately.

Return type

Model{T, Triangle{T}}

Alias

readmsms{T}(fname::String, ::Type{T}=Float64)

Reads the model using a common file name prefix (fname.{vert,face}) for both files rather than IOStream objects.

source

OFF

NESSie.Format.readoff — Function.

readoff{T <: AbstractFloat}(
    stream::IOStream,
          ::Type{T}=Float64
)

Reads a surface model from the given OFF file.

Note

This file type does not support charge models! Hence, the charge list of the returning Model object is empty and has to be set separately.

Specification

http://www.geomview.org/docs/html/OFF.html

Return type

Model{T}

Alias

readoff{T}(fname::String, ::Type{T}=Float64)

Reads the model using a file name rather than a IOStream object.

source

PQR

NESSie.Format.readpqr — Function.

readpqr{T <: AbstractFloat}(
    stream::IOStream,
          ::Type{T}=Float64
)

Reads a charge model from the given PQR file.

Return type

Vector{Charge{T}}

Alias

readpqr{T}(fname::String, ::Type{T}=Float64)

Reads the charge model using a file name rather than a IOStream object.

source