Output formats

Output formats

Currently supported output file formats with different models:

File type	Point cloud	Surface model	Volume model

SKEL		✓	✓

VTK		✓	✓

XML3D/JSON	✓	✓

XML3D/XML	✓

SKEL

NESSie.Format.writeskel — Function.

writeskel{T, M <: Model{T}}(
    stream::IOStream,
    model ::M
)

Creates a SKEL file from a given surface or volume model, representing the model as a collection of points and polylines.

Specification

http://www.geomview.org/docs/html/SKEL.html

Return type

Void

Alias

writeskel{T, M <: Model{T}}(
    fname::String,
    model::M
)

Creates the SKEL file by name rather than IOStream object.

source

VTK

NESSie.Format.writevtk — Function.

function writevtk{T, M <: Model{T}}(
    stream::IOStream,
    model ::M
)

Creates a VTK-compatible output file from a given surface or volume model. The exact file type is determined by the given model:

Model type	Resulting file type

Surface model	VTK PolyData

Volume model	VTK UnstructuredGrid

Specification

http://www.vtk.org/VTK/img/file-formats.pdf

Return type

Void

Alias

writevtk{T, M <: Model{T}}(
    fname::String,
    model::M
)

Creates the VTK file by name rather than IOStream object.

source

XML3D

NESSie.Format.writexml3d_json — Function.

writexml3d_json{T}(
    stream::IOStream,
    model ::Union{Vector{Vector{T}}, Model{T, Triangle{T}}}
)

Creates a XML3D-specific JSON file either from a given collection of nodes (representing the latter as point cloud) or from a given surface model.

Specification

https://github.com/xml3d/xml3d.js/wiki/External-resources

Return type

Void

Alias

writexml3d_json{T}(
    fname::String,
    nodes::Union{Vector{Vector{T}}), Model{T, Triangle{T}}}
)

Creates the JSON file by name rather than IOStream object.

source

NESSie.Format.writexml3d_xml — Function.

writexml3d_xml{T}(
    stream::IOStream,
    nodes ::Vector{Vector{T}}
)

Creates a XML3D-specific XML file from a given collection of nodes, representing the latter as point cloud.

Specification

https://github.com/xml3d/xml3d.js/wiki/External-resources

Return type

Void

Alias

writexml3d_xml{T}(
    fname::String,
    nodes::Vector{Vector{T}}
)

Creates the XML file by name rather than IOStream object.

source