NESSie.BEM

NESSie.BEM.Kfun

NESSie.BEM.Kyfun

NESSie.BEM.LocalSystemMatrix

NESSie.BEM.NonlocalSystemMatrix

NESSie.BEM.Vfun

NESSie.BEM.Vyfun

NESSie.BEM._espotential_Γ

NESSie.BEM._espotential_Σ

NESSie.BEM._espotential_Ω

NESSie.BEM._get_laplace_matrices

NESSie.BEM._get_yukawa_matrices

NESSie.BEM._rfpotential_Γ

NESSie.BEM._rfpotential_Σ

NESSie.BEM._rfpotential_Ω

NESSie.BEM._solve_explicit

NESSie.BEM._solve_implicit

NESSie.BEM._solve_linear_system

Solvers

NESSie.BEM._get_laplace_matrices — Function

_get_laplace_matrices(
    Ξ       ::Vector{Vector{T}},
    elements::Vector{Triangle{T}}
)

Constructs and returns implicit representations for the single- and double-layer Laplace potential matrices V and K for the given observation points and surface elements.

Return type

Tuple{InteractionMatrix{T}, InteractionMatrix{T}}

NESSie.BEM._get_yukawa_matrices — Function

_get_yukawa_matrices(
    Ξ       ::Vector{Vector{T}},
    elements::Vector{Triangle{T}},
    yuk     ::T
)

Constructs and returns implicit representations for the single- and double-layer Yukawa potential matrices V^Y and K^Y for the given observation points and surface elements.

Return type

Tuple{InteractionMatrix{T}, InteractionMatrix{T}}

NESSie.BEM._solve_explicit — Function

_solve_explicit(
         ::Type{<: LocalityType},
    model::Model{T, Triangle{T}}
)

Computes the full local or nonlocal cauchy data on the surface of the biomolecule using an explicit representation of the BEM system.

Return type

LocalBEMResult{T, Triangle{T}} or NonlocalBEMResult{T, Triangle{T}}

NESSie.BEM._solve_implicit — Function

_solve_implicit(
         ::Type{<: LocalityType},
    model::Model{T, Triangle{T}}
)

Computes the full local or nonlocal cauchy data on the surface of the biomolecule using an implicit representation of the BEM system.

Return type

LocalBEMResult{T, Triangle{T}} or NonlocalBEMResult{T, Triangle{T}}

NESSie.BEM._solve_linear_system — Function

_solve_linear_system(
    A::AbstractArray{T, 2},
    b::AbstractArray{T, 1}
)

Solves the linear system Ax = b from A and b using a Jacobi-preconditioned GMRES solver.

Return type

Array{T, 1}

Supported keyword arguments

Everything from IterativeSolvers.gmres except for log.

Implicit potential matrices

NESSie.BEM.Kfun — Type

struct Kfun{T} <: InteractionFunction{Vector{T}, Triangle{T}, T} end

Interaction function for an implicit representation of potential matrix K.

NESSie.BEM.Kyfun — Type

struct Kyfun{T} <: InteractionFunction{Vector{T}, Triangle{T}, T}
    yuk::T   # exponent of the Yukawa operator's fundamental solution
end

Interaction function for an implicit representation of potential matrix Kʸ.

NESSie.BEM.Vfun — Type

struct Vfun{T} <: InteractionFunction{Vector{T}, Triangle{T}, T} end

Interaction function for an implicit representation of potential matrix V.

NESSie.BEM.Vyfun — Type

struct Vyfun{T} <: InteractionFunction{Vector{T}, Triangle{T}, T}
    yuk::T   # exponent of the Yukawa operator's fundamental solution
end

Interaction function for an implicit representation of potential matrix Vʸ.

Implicit system matrices

NESSie.BEM.LocalSystemMatrix — Type

struct LocalSystemMatrix{T} <: AbstractArray{T, 2}
    K     ::InteractionMatrix{T}   # double-layer Laplace
    params::Option{T}              # System constants
end

Implicit representation of the first local BEM system matrix.

NESSie.BEM.NonlocalSystemMatrix — Type

struct NonlocalSystemMatrix{T} <: AbstractArray{T, 2}
    V     ::InteractionMatrix{T}   # single-layer Laplace
    K     ::InteractionMatrix{T}   # double-layer Laplace
    Vy    ::InteractionMatrix{T}   # single-layer Yukawa
    Ky    ::InteractionMatrix{T}   # double-layer Yukawa
    params::Option{T}              # system constants
end

Implicit representation of the nonlocal BEM system matrix.

Special constructors

function NonlocalSystemMatrix{T}(    Ξ       ::Vector{Vector{T}},    elements::Vector{Triangle{T}},    params  ::Option{T}) where T

Creates the nonlocal system matrix directly from the given observation points Ξ and surface elements.

Potentials

NESSie.BEM._espotential_Γ — Function

_espotential_Γ(ξ::Vector{T}, bem::BEMResult{T})
_espotential_Γ(Ξ::AbstractArray{Vector{T}}, bem::BEMResult{T})

Computes the local or nonlocal electrostatic potential for (an) observation point(s) ξ (Ξ) on the molecular surface.

Supported keyword arguments

See molpotential

Unit

$V = \frac{C}{F}$

Return type

T or Array{T}

NESSie.BEM._espotential_Ω — Function

_espotential_Ω(ξ::Vector{T}, bem::BEMResult{T})
_espotential_Ω(Ξ::AbstractArray{Vector{T}}, bem::BEMResult{T})

Computes the local or nonlocal electrostatic potential for (an) observation point(s) ξ (Ξ) inside the molecule.

Supported keyword arguments

See molpotential

Unit

$V = \frac{C}{F}$

Return type

T or Array{T}

NESSie.BEM._espotential_Σ — Function

_espotential_Σ(ξ::Vector{T}, bem::BEMResult{T})
_espotential_Σ(Ξ::AbstractArray{Vector{T}}, bem::BEMResult{T})

Computes the local or nonlocal electrostatic potential for (an) observation point(s) ξ (Ξ) in the solvent domain.

Unit

$V = \frac{C}{F}$

Return type

T or Array{T}

NESSie.BEM._rfpotential_Γ — Function

_rfpotential_Γ(ξ::Vector{T}, bem::BEMResult{T})
_rfpotential_Γ(Ξ::AbstractArray{Vector{T}}, bem::BEMResult{T})

Computes the local or nonlocal reaction field potential for (an) observation point(s) ξ (Ξ) on the molecular surface.

Unit

$V = \frac{C}{F}$

Return type

T or Array{T}

NESSie.BEM._rfpotential_Ω — Function

_rfpotential_Ω(ξ::Vector{T}, bem::BEMResult{T})
_rfpotential_Ω(Ξ::AbstractArray{Vector{T}}, bem::BEMResult{T})

Computes the local or nonlocal reaction field potential for (an) observation point(s) ξ (Ξ) inside the molecule.

Unit

$V = \frac{C}{F}$

Return type

T or Array{T}

NESSie.BEM._rfpotential_Σ — Function

_rfpotential_Σ(ξ::Vector{T}, bem::BEMResult{T})
_rfpotential_Σ(Ξ::AbstractArray{Vector{T}}, bem::BEMResult{T})

Computes the local or nonlocal reaction field potential for (an) observation point(s) ξ (Ξ) in the solvent domain.

Supported keyword arguments

See molpotential

Unit

$V = \frac{C}{F}$

Return type

T or Array{T}

mutable struct Option{T <: AbstractFloat} εΩ::T # dielectric constant of the solute εΣ::T # dielectric constant of the solvent ε∞::T # high-frequency solvent response λ ::T # correlation length scale [λ] = Å end
System parameters

abstract type LocalityType end struct NonlocalES <: LocalityType end struct LocalES <: LocalityType end
Enum-like representation of locality assumptions

mutable struct Model{T, E <: Element{T}} nodes ::Vector{Vector{T} = Vector{T}[] # mesh nodes elements::Vector{E} = E[] # mesh elements charges ::Vector{Charge{T}} = Charge{T}[] # point charges in the molecule params ::Option{T} = defaultopt(T) # system constants end
System model representing a biomelecule in solvation, including a collection of point charges in the molecule and a set of system constants.

struct Triangle{T} <: SurfaceElement{T} v1 ::Vector{T} # position of the first node v2 ::Vector{T} # position of the second node v3 ::Vector{T} # position of the third node center ::Vector{T} # centroid of the triangle normal ::Vector{T} # normal vector of the triangle area ::T # area of the triangle distorig::T # distance to the origin end
Representation of a single surface triangle.

abstract type BEMResult{T, E <: SurfaceElement{T}} end
Abstract base type for all BEM solver results