Quadrature

Quadrature

Radon cubature

NESSie.Radon.radoncoll! — Function.

radoncoll!{T}(
        dest    ::DenseArray{T,1},
        elements::Vector{Triangle{T}},
        Ξ       ::Vector{Vector{T}},
        solution::Function,
        yukawa  ::T,
        fvals   ::DenseArray{T,1}
)

radoncoll!{T}(
        dest    ::DenseArray{T,2},
        elements::Vector{Triangle{T}},
        Ξ       ::Vector{Vector{T}},
        solution::Function,
        yukawa  ::T
)

Seven-point Radon cubature [Rad48] for a given function and a list of triangles and observation points Ξ. If dest is a vector, the function values f for each surface triangle have to be specified, since each element of the vector represents the dot product of the corresponding coefficient matrix row and the fvals vector.

If you intend computing single/double layer potentials with this function, you might want to use the shorthand signature regularyukawacoll! instead.

Note

The result is premultiplied by 4π.

Arguments

solution Fundamental solution; supported functions: regularyukawapot, ∂ₙregularyukawapot

yukawa Exponent of the Yukawa operator's fundamental solution

Return type

Void

source

NESSie.Radon.regularyukawapot — Function.

regularyukawapot{T}(
    x     ::DenseArray{T,1},
    ξ     ::Vector{T},
    yukawa::T,
          ::Vector{T}=T[]
)

Computes the regular part of the Yukawa potential, that is, Yukawa minus Laplace:

\[\frac{e^{Λ |x - ξ|} - 1}{|x-ξ|}\]

Note

The result is premultiplied by 4π.

Arguments

yukawa Exponent of the Yukawa operator's fundamental solution

Return type

T

source

NESSie.Radon.∂ₙregularyukawapot — Function.

∂ₙregularyukawapot{T}(
    x     ::Vector{T},
    ξ     ::Vector{T},
    yukawa::T,
    normal::Vector{T}
)

Computes the normal derivative of the regular part of the Yukawa potential, that is, Yukawa minus Laplace:

\[\frac{∂}{∂n} \frac{e^{Λ |x - ξ|} - 1}{|x-ξ|} = \frac{1 - (1 - Λ |x - ξ|)e^{Λ |x - ξ|}}{|x-ξ|²} \frac{(x - ξ) ⋅ n}{|x - ξ|}\]

Note

The result is premultiplied by 4π.

Arguments

yukawa Exponent of the Yukawa operator's fundamental solution

Return type

T

source

NESSie.Radon.setcubpts! — Function.

setcubpts!{T}(
    dest::Vector{Vector{T}},
    qpts::QuadPts2D{T},
    elem::Triangle{T}
)

Prepare cubature points for one surface element.

Return type

Void

source

Rjasanow

NESSie.Rjasanow.ObservationPosition — Type.

abstract type ObservationPosition end
struct InPlane <: ObservationPosition end
stryct InSpace <: ObservationPosition end

Enum-like representation of the obseration point's position relative to the corresponding surface element

source

NESSie.Rjasanow.InPlane — Type.

abstract type ObservationPosition end
struct InPlane <: ObservationPosition end
stryct InSpace <: ObservationPosition end

Enum-like representation of the obseration point's position relative to the corresponding surface element

source

NESSie.Rjasanow.InSpace — Type.

abstract type ObservationPosition end
struct InPlane <: ObservationPosition end
stryct InSpace <: ObservationPosition end

Enum-like representation of the obseration point's position relative to the corresponding surface element

source

NESSie.Rjasanow.laplacepot — Function.

function laplacepot{T, P <: PotentialType}(
    ptype::Type{P},
    ξ    ::Vector{T},
    elem ::Triangle{T},
    dist ::T
)

Computes the single or double layer Laplace potential of the given triangle for the given observation point ξ. The latter needs to be projected onto the surface element plane [Rja90].

Note

The result is premultiplied by 4π.

Arguments

dist Distance from the original ξ to the surface element plane

Return type

T

source

laplacepot{T, P <: PotentialType}(
    ptype::Type{P},
    ξ::Vector{T},
    x1::Vector{T},
    x2::Vector{T},
    normal::Vector{T},
    dist::T
)

Computes the single or double layer Laplace potential of the triangle defined by the observation point ξ and two nodes x1 and x2 of the surface element. x2 is required to be x1's next neighbor in counterclockwise direction. Also, ξ needs to be projected onto the surface element plane [Rja90].

Note

The result is premultiplied by 4π.

Arguments

normal Unit normal vector of the surface element

dist Distance from the original ξ to the surface element plane

Return type

T

source

laplacepot{T, P <: PotentialType, O <: ObservationPosition}(
         ::Type{P},
         ::Type{O},
    sinφ1::T,
    sinφ2::T,
    h    ::T,
    d    ::T
)

Computes the Laplace potential (or its normal derivative) of the triangle with the given height h at the observation point ξ (projected onto the surface element plane) and the sines of the angles $φ₁$ and $φ₂$ between h and the triangle sides extending from ξ [Rja90].

Note

The result is premultiplied by 4π.

Arguments

d Distance from the original ξ to the surface element plane

Return type

T

source

NESSie.Rjasanow.logterm — Function.

logterm{T}(χ2::T, sinφ::T)

Utility function to compute

\[\frac {\sqrt{1 - χ² \sin²(φ)} + \sqrt{1 - χ² \sin(φ)}} {\sqrt{1 - χ² \sin²(φ)} - \sqrt{1 - χ² \sin(φ)}}\]

Return type

T

source

NESSie.Rjasanow.projectξ — Function.

projectξ{T}(ξ::Vector{T}, elem::Triangle{T})

Projects ξ onto the surface element plane. Returns the projection and the original distance to the plane.

Return type

Tuple{Vector{T}, T}

source