Electrostatics

Electrostatics

Potential types

NESSie.PotentialType — Type.

abstract type PotentialType end
struct SingleLayer <: PotentialType end
struct DoubleLayer <: PotentialType end

Enum-like representation of single and double layer potentials

source

NESSie.SingleLayer — Type.

abstract type PotentialType end
struct SingleLayer <: PotentialType end
struct DoubleLayer <: PotentialType end

Enum-like representation of single and double layer potentials

source

NESSie.DoubleLayer — Type.

abstract type PotentialType end
struct SingleLayer <: PotentialType end
struct DoubleLayer <: PotentialType end

Enum-like representation of single and double layer potentials

source

Locality assumption

NESSie.LocalityType — Type.

abstract type LocalityType end
struct NonlocalES <: LocalityType end
struct LocalES    <: LocalityType end

Enum-like representation of locality assumption:

Local electrostatics: Complete independence of solvent molecules

Nonlocal electrostatics: Allow solvent molecule correlation effects (with area-of-effect radius λ)

source

NESSie.LocalES — Type.

abstract type LocalityType end
struct NonlocalES <: LocalityType end
struct LocalES    <: LocalityType end

Enum-like representation of locality assumption:

Local electrostatics: Complete independence of solvent molecules

Nonlocal electrostatics: Allow solvent molecule correlation effects (with area-of-effect radius λ)

source

NESSie.NonlocalES — Type.

abstract type LocalityType end
struct NonlocalES <: LocalityType end
struct LocalES    <: LocalityType end

Enum-like representation of locality assumption:

Local electrostatics: Complete independence of solvent molecules

Nonlocal electrostatics: Allow solvent molecule correlation effects (with area-of-effect radius λ)

source

Potentials

Molecular potentials

NESSie.φmol — Function.

φmol{T}(ξ::Vector{T}, charges::Vector{Charge{T}})

Computes and returns the molecular potential of the given system of point charges in a structureless medium for the given observation point ξ:

\[φ_{mol}(ξ) = \frac{1}{4π ε_0 ε_Ω} \sum_i \frac{qᵢ}{|rᵢ-ξ|}\]

Note

The return value is premultiplied by $4π ⋅ ε₀ ⋅ ε_Ω$

Return type

T

Aliases

φmol{T}(Ξ::Vector{Vector{T}}, charges::Vector{Charge{T}})

Computes the molecular potentials for a list of observation points.

φmol{T}(model::Model{T, Triangle{T}})

Computes the molecular potentials for the given surface model, using each triangle center as observation point.

source

NESSie.∂ₙφmol — Function.

∂ₙφmol{T}(ξ::Triangle{T}, charges::Vector{Charge{T}})

Computes and returns the normal derivative of the given system's molecular potential in a structureless medium, using the given triangle's center as observation point and the triangle's normal as reference normal.

\[∂ₙφ_{mol}(ξ) = -\frac{1}{4π ε_0 ε_Ω} \sum_i \frac{qᵢ}{|rᵢ-ξ|³} (rᵢ-ξ) ⋅ n\]

Note

The return value is premultiplied by $4π ⋅ ε₀ ⋅ ε_Ω$

Return type

T

Aliases

∂ₙφmol{T}(model::Model{T, Triangle{T}})

Computes the normal derivatives of the molecular potentials for the given surface model, using each triangle center and normal as observation point.

source

NESSie.∇φmol — Function.

∇φmol{T}(ξ::Vector{T}, charges::Vector{Charge{T}})

Computes and returns the gradient of the given system's molecular potential in a structureless medium for the given observation point ξ.

\[∇φ_{mol}(ξ) = -\frac{1}{4π ε_0 ε_Ω} \sum_i \frac{qᵢ}{|rᵢ-ξ|³} (rᵢ-ξ)\]

Note

The return value is premultiplied by $4π ⋅ ε₀ ⋅ ε_Ω$

Return type

Vector{T}

Aliases

∇φmol{T}(Ξ::Vector{Vector{T}})

Computes the molecular potential gradients for a list of observation points.

source

Interior potentials

NESSie.BEM.φΩ — Function.

φΩ{T, B <: BEMResult{T}}(
    Ξ         ::Vector{Vector{T}},
    bem       ::B
)

Computes the local or nonlocal interior electrostatic potential $φ_Ω$ for the given set of observation points Ξ.

Warning

This function does not verify whether all points in Ξ are located in $Ω$!

Unit

$V = \frac{C}{F}$

Return type

Vector{T}

source

NESSie.Born.φΩ — Function.

φΩ{T, L <: LocalityType}(
       ::Type{L},
    ξ  ::Vector{T},
    ion::BornIon{T},
    opt::Option{T} = defaultopt(T)
)

Computes the interior local or nonlocal electrostatic potential $φ_Ω$ for the given observation point $ξ$.

Unit

$V = \frac{C}{F}$

Return type

T

Warning

This function does not verify whether ξ is located inside of the sphere!

source

Exterior potentials

NESSie.BEM.φΣ — Function.

φΣ{T, B <: BEMResult{T}}(
    Ξ         ::Vector{Vector{T}},
    bem       ::B
)

Computes the local or nonlocal exterior electrostatic potential $φ_Σ$ for the given set of observation points Ξ.

Warning

This function does not verify whether all points in Ξ are located in $Σ$!

Unit

$V = \frac{C}{F}$

Return type

Vector{T}

source

NESSie.Born.φΣ — Function.

φΣ{T, L <: LocalityType}(
       ::Type{L},
    ξ  ::Vector{T},
    ion::BornIon{T},
    opt::Option{T} = defaultopt(T)
)

Computes the exterior local or nonlocal electrostatic potential $φ_Σ$ for the given observation point $ξ$.

Unit

$V = \frac{C}{F}$

Return type

T

Warning

This function does not verify whether ξ is located outside of the sphere!

source

Energies

NESSie.BEM.rfenergy — Function.

rfenergy{T, B <: BEMResult{T}}(
    bem       ::B
)

Computes the local or nonlocal reaction field energy W* as

\[W^* = ∫φ^* ρ \quad dΩ\]

where $φ^*$ is the reaction field and $ρ$ is the corresponding charge distribution.

Unit

$\frac{kJ}{mol}$

Return type

T

source