Models

Models

Elements

NESSie.Element — Type.

abstract type Element{T <: AbstractFloat} end
abstract type SurfaceElement{T} <: Element{T} end
abstract type VolumeElement{T}  <: Element{T} end

Abstract base types for all elements.

source

NESSie.Triangle — Type.

struct Triangle{T} <: SurfaceElement{T}
    v2      ::Vector{T}   # position of the second node
    v1      ::Vector{T}   # position of the first node
    v3      ::Vector{T}   # position of the third node
    center  ::Vector{T}   # centroid of the triangle
    normal  ::Vector{T}   # normal vector of the triangle
    area    ::T           # area of the triangle
    distorig::T           # distance to the origin
end

Representation of a single surface triangle.

Special constructors

Triangle{T}(
    v1  ::Vector{T},
    v2  ::Vector{T},
    v3  ::Vector{T}
)

Most commonly used constructor variant for creating a triangle by only specifying its nodes. The remaining member variables will automatically be computed via props.

source

NESSie.Tetrahedron — Type.

struct Tetrahedron{T} <: VolumeElement{T}
    v1::Vector{T}       # position of the first node
    v2::Vector{T}       # position of the second node
    v3::Vector{T}       # position of the third node
    v4::Vector{T}       # position of the fourth node
    domain::Symbol      # element domain (solvent :Σ, solute :Ω, or :none)
end

Representation of a single tetrahedron.

Special constructors

Tetrahedron{T}(
    v1::Vector{T},
    v2::Vector{T},
    v3::Vector{T},
    v4::Vector{T}
)

Sets domain to :none.

source

Charge models

NESSie.Charge — Type.

struct Charge{T <: AbstractFloat}
    pos::Vector{T}  # position of the charge
    val::T          # charge value
end

Representation of a single point charge.

Special constructors

Charge{T}(
    posx::T,
    posy::T,
    posz::T,
    val ::T
)

Constructor variant with flat argument list for pos.

source

NESSie.Born.BornIon — Type.

struct BornIon{T <: AbstractFloat}
    charge::Charge{T}  # point charge at the sphere's center
    radius::T          # sphere radius in Å
end

Single Born ion, that is, a monoatomic ion represented as a spherically symmetric domain with a single point charge located at its center ($ε_Ω = 1$).

Special constructors

BornIon{T}(charge::T, radius::T)

Centers the sphere at $(0, 0, 0)^T$.

source

System models

NESSie.Model — Type.

mutable struct Model{T, E <: Element{T}}
    nodes   ::Vector{Vector{T}  = Vector{T}[]    # mesh nodes
    elements::Vector{E}         = E[]            # mesh elements
    charges ::Vector{Charge{T}} = Charge{T}[]    # point charges in the molecule
    params  ::Option{T}         = defaultopt(T)  # system constants
end

System model representing a biomelecule in solvation, including a collection of point charges in the molecule and a set of system constants. The system can either be represented as a surface model (e.g., a collection of molecule surface triangles) or as a volume model (e.g., a collection of tetrahedra for the molecule and its surrounding space).

source