Quadrature

Quadrature

Quadrature points

NESSie.QuadraturePoints — Type.

abstract type QuadraturePoints{T <: AbstractFloat} end

Abstract base type for all sets of quadrature points

source

NESSie.QuadPts2D — Type.

struct QuadPts2D{T} <: QuadraturePoints{T}
    num   ::Int         # number of points
    x     ::Vector{T}   # x values
    y     ::Vector{T}   # y values
    weight::Vector{T}   # weights
end

Quadrature points and weights for triangles

source

NESSie.QuadPts3D — Type.

struct QuadPts3D{T} <: QuadraturePoints{T}
    num   ::Int         # number of points
    x     ::Vector{T}   # x values
    y     ::Vector{T}   # y values
    z     ::Vector{T}   # z values
    weight::Vector{T}   # weights
end

Quadrature points and weights for tetrahedra

source

Generators

NESSie.quadraturepoints — Function.

quadraturepoints(::Type{Tetrahedron{Float64}})
quadraturepoints(::Type{Tetrahedron{Float32}})
quadraturepoints(::Type{Triangle{Float64}})
quadraturepoints(::Type{Triangle{Float32}})

Generator function for quadrature points:

Triangles: 7 points per element [Rad48]

Tetrahedra: 5 points per element [Kea86]

Return type

QuadPts2D or QuadPts3D

source

Laplace potential

NESSie.Rjasanow.laplacecoll! — Function.

laplacecoll!{T, P <: PotentialType}(
    ptype   ::Type{P},
    dest    ::DenseArray{T,1},
    elements::Vector{Triangle{T}},
    Ξ       ::Vector{Vector{T}},
    fvals   ::Union{DenseArray{T,1},SubArray{T,1}}
)

laplacecoll!{T, P <: PotentialType}(
    ptype   ::Type{P},
    dest    ::DenseArray{T, 2},
    elements::Vector{Triangle{T}},
    Ξ       ::Vector{Vector{T}}
)

Analytical solution for the single or double layer Laplace potential for a given list of triangles and observation points Ξ [Rja90].

The first version of this function uses a vector as destination dest, where each element represents the dot product of the corresponding coefficient matrix row and the fvals vector.

Note

The result is premultiplied by 4π.

Return type

Void

source

Yukawa potential

NESSie.Radon.regularyukawacoll! — Function.

regularyukawacoll!{T, P <: PotentialType}(
            ::Type{P},
    dest    ::DenseArray{T,1},
    elements::Vector{Triangle{T}},
    Ξ       ::Vector{Vector{T}},
    yukawa  ::T,
    fvals   ::Union{DenseArray{T,1},SubArray{T,1}}
)

regularyukawacoll!{T, P <: PotentialType}(
            ::Type{P},
    dest    ::DenseArray{T,2},
    elements::Vector{Triangle{T}},
    Ξ       ::Vector{Vector{T}},
    yukawa  ::T
)

Computes the regular part of the single or double layer Yukawa potential (that is, Yukawa minus Laplace) using a seven-point Radon cubature [Rad48] for a given list of triangles and observation points Ξ.

The first version of this function uses a vector as destination dest, where each element represents the dot product of the corresponding coefficient matrix row and the fvals vector.

Note

The result is premultiplied by 4π.

Arguments

yukawa Exponent of the Yukawa operator's fundamental solution

Return type

Void

source