Solvers

Solvers

BEM solver

NESSie.BEM.BEMResult — Type.

abstract type BEMResult{T, E <: SurfaceElement{T}} end

Abstract base type for all BEM solver results

source

NESSie.BEM.LocalBEMResult — Type.

struct LocalBEMResult{T, E} <: BEMResult{T, E}
    model::Model{T, E}
    u    ::Vector{T}   # [γ₀int(φ*)](ξ)    ∀ ξ ∈ Ξ; premultiplied by 4π⋅ε0
    q    ::Vector{T}   # [γ₁int(φ*)](ξ)    ∀ ξ ∈ Ξ; premultiplied by 4π⋅ε0
    umol ::Vector{T}   # [γ₀int(φ*mol)](ξ) ∀ ξ ∈ Ξ; premultiplied by 4π⋅ε0
    qmol ::Vector{T}   # [γ₁int(φ*mol)](ξ) ∀ ξ ∈ Ξ; premultiplied by 4π⋅ε0
end

Result data of the local solving process to be used for potential computation and post-processing, with Ξ being the list of observation points, that is, the set of triangle centroids.

source

NESSie.BEM.NonlocalBEMResult — Type.

struct NonlocalBEMResult{T, E} <: BEMResult{T, E}
    model::Model{T, E}
    u    ::SubArray{T,1}   # [γ₀int(φ*)](ξ)    ∀ ξ ∈ Ξ; premultiplied by 4π⋅ε0
    q    ::SubArray{T,1}   # [γ₁int(φ*)](ξ)    ∀ ξ ∈ Ξ; premultiplied by 4π⋅ε0
    w    ::SubArray{T,1}   # [γ₀ext(Ψ)](ξ)     ∀ ξ ∈ Ξ; premultiplied by 4π⋅ε0
    umol ::Vector{T}       # [γ₀int(φ*mol)](ξ) ∀ ξ ∈ Ξ; premultiplied by 4π⋅ε0
    qmol ::Vector{T}       # [γ₁int(φ*mol)](ξ) ∀ ξ ∈ Ξ; premultiplied by 4π⋅ε0
end

Result data of the nonlocal solving process to be used for potential computation and post-processing, with Ξ being the list of observation points, that is, the set of triangle centroids.

source

NESSie.BEM.solve — Function.

solve{T, L <: LocalityType}(
              ::L,
    model     ::Model{T, Triangle{T}}
)

Computes the full local or nonlocal cauchy data on the surface of the biomolecule.

Return type

LocalBEMResult{T, Triangle{T}} or NonlocalBEMResult{T, Triangle{T}}

source