Utility functions

Utility functions

NESSie.Born.bornion — Function.

bornion(name::String, ::Type{Float64} = Float64)
bornion(name::String, ::Type{Float32})

Generator function for built-in Born ions:

Name	Charge	Radius [Åqv90]

Li	+1	0.645

Na	+1	1.005

K	+1	1.365

Rb	+1	1.505

Cs	+1	1.715

Mg	+2	0.615

Ca	+2	1.015

Sr	+2	1.195

Ba	+2	1.385

Return type

BornIon

source

NESSie.meshunion — Function.

meshunion{T}(
    model1::Model{T, Tetrahedron{T}},
    model2::Model{T, Tetrahedron{T}}
)

Merges two volume models, e.g., the models of a protein and the solvent. Duplicate nodes (e.g., the nodes on the protein surface) are merged into a single node, duplicate elements and charges (if any) are retained as well as the system constants of the first model.

Return type

Model{T, Tetrahedron{T}}

Note

This function assumes that there are no duplicates within either of the node lists!

source

NESSie.obspoints_line — Function.

obspoints_line{T}(
    u::Vector{T},
    v::Vector{T},
    n::Int
)

Generates n evenly distributed observation points along the line segment from u to v.

Return type

Function

Example

for ξ in obspoints_line([0, 0, 0], [1, 1, 1], 10)
    ...
end

source

NESSie.obspoints_plane — Function.

obspoints_plane{T}(
    a  ::Vector{T},
    b  ::Vector{T},
    c  ::Vector{T},
    nba::Int,
    nbc::Int
)

Generates nba ⋅ nbc evenly distributed observation points on the parallelogram with the sides $\overline{BA}$ and $\overline{BC}$ and a, b, c being the location vectors to the points $A$, $B$, and $C$, respectively.

Arguments

nba Number of observation points along $\overline{BA}$

nbc Number of observation points along $\overline{BC}$

Return type

Function

Example

for Ξ in obspoints_plane(...)
    for ξ in Ξ
        ...
    end
end

source