Output formats

Currently supported output file formats with different models:

File type	Point cloud	Surface model	Volume model	Charges included

HMO		✓		✓

OBJ		✓

SKEL		✓	✓

STL		✓

VTK		✓	✓

XML3D/JSON	✓	✓

XML3D/XML	✓

HMO

NESSie.Format.writehmo — Function

writehmo{T}(
    stream::IOStream,
    model ::Model{T, Triangle{T}}
)

Creates a HMO file from the given surface model.

Return type

Void

Alias

writehmo{T}(
    fname::String,
    model::Model{T, Triangle{T}}
)

Creates the HMO file by name rather than IOStream object.

OBJ

NESSie.Format.writeobj — Function

function writeobj{T}(
    stream::IOStream,
    model ::Model{T, Triangle{T}}
)

Creates a Wavefront OBJ file from a given surface model.

Specification

<https://www.loc.gov/preservation/digital/formats/fdd/fdd000507.shtml>
<http://paulbourke.net/dataformats/obj/>

Return type

Void

Alias

writeobj{T}(
    fname::String,
    model::Model{T, Triangle{T}}
)

Creates the OBJ file by name rather than IOStream object.

SKEL

NESSie.Format.writeskel — Function

writeskel{T, M <: Model{T}}(
    stream::IOStream,
    model ::M
)

Creates a SKEL file from a given surface or volume model, representing the model as a collection of points and polylines.

Specification

http://www.geomview.org/docs/html/SKEL.html

Return type

Void

Alias

writeskel{T, M <: Model{T}}(
    fname::String,
    model::M
)

Creates the SKEL file by name rather than IOStream object.

STL

NESSie.Format.writestl — Function

writestl{T}(
    stream::IOStream,
    model ::Model{T, Triangle{T}}
)

Creates a binary STL file from a given surface model.

Specification

http://www.fabbers.com/tech/STL_Format

Return type

Void

Alias

writestl{T}(
    fname::String,
    model::Model{T, Triangle{T}}
)

Creates the STL file by name rather than IOStream object.

VTK

NESSie.Format.writevtk — Function

writevtk{T, M <: Model{T}}(
    stream::IOStream,
    model ::M
)

Creates a VTK-compatible output file from a given surface or volume model. The exact file type is determined by the given model:

Model type	Resulting file type

Surface model	VTK PolyData

Volume model	VTK UnstructuredGrid

Specification

https://vtk.org/wp-content/uploads/2015/04/file-formats.pdf

Return type

Void

Alias

writevtk{T, M <: Model{T}}(
    fname::String,
    model::M
)

Creates the VTK file by name rather than IOStream object.

XML3D

NESSie.Format.writexml3d_json — Function

writexml3d_json{T}(
    stream::IOStream,
    model ::Union{Vector{Vector{T}}, Model{T, Triangle{T}}}
)

Creates a XML3D-specific JSON file either from a given collection of nodes (representing the latter as point cloud) or from a given surface model.

Specification

https://github.com/xml3d/xml3d.js/wiki/External-resources

Return type

Void

Alias

writexml3d_json{T}(
    fname::String,
    nodes::Union{Vector{Vector{T}}), Model{T, Triangle{T}}}
)

Creates the JSON file by name rather than IOStream object.

NESSie.Format.writexml3d_xml — Function

writexml3d_xml{T}(
    stream::IOStream,
    nodes ::Vector{Vector{T}}
)

Creates a XML3D-specific XML file from a given collection of nodes, representing the latter as point cloud.

Specification

https://github.com/xml3d/xml3d.js/wiki/External-resources

Return type

Void

Alias

writexml3d_xml{T}(
    fname::String,
    nodes::Vector{Vector{T}}
)

Creates the XML file by name rather than IOStream object.