NESSie.BEM

Implicit potential matrices

NESSie.BEM.Kfun — Type

struct Kfun{T} <: InteractionFunction{Vector{T}, Triangle{T}, T} 
    dat::Vector{T}   # pre-allocated vector for internal use
end

Interaction function for an implicit representation of potential matrix K.

Special constructors

Kfun{T}()

Automatically initializes the internal data vector. Replaces the default constructor.

NESSie.BEM.Kyfun — Type

struct Kyfun{T} <: InteractionFunction{Vector{T}, Triangle{T}, T}
    yuk::T   # exponent of the Yukawa operator's fundamental solution
end

Interaction function for an implicit representation of potential matrix Kʸ.

NESSie.BEM.Vfun — Type

struct Vfun{T} <: InteractionFunction{Vector{T}, Triangle{T}, T}
    dat::Vector{T}   # pre-allocated vector for internal use
end

Interaction function for an implicit representation of potential matrix V.

Special constructors

Vfun{T}()

Automatically initializes the internal data vector. Replaces the default constructor.

NESSie.BEM.Vyfun — Type

struct Vyfun{T} <: InteractionFunction{Vector{T}, Triangle{T}, T}
    yuk::T   # exponent of the Yukawa operator's fundamental solution
end

Interaction function for an implicit representation of potential matrix Vʸ.

Implicit system matrices

NESSie.BEM.LocalSystemMatrix — Type

struct LocalSystemMatrix{T} <: AbstractArray{T, 2}
    K     ::InteractionMatrix{T}   # double-layer Laplace
    params::Option{T}              # System constants
end

Implicit representation of the first local BEM system matrix.

NESSie.BEM.NonlocalSystemMatrix — Type

struct NonlocalSystemMatrix{T} <: AbstractArray{T, 2}
    V     ::InteractionMatrix{T}   # single-layer Laplace
    K     ::InteractionMatrix{T}   # double-layer Laplace
    Vy    ::InteractionMatrix{T}   # single-layer Yukawa
    Ky    ::InteractionMatrix{T}   # double-layer Yukawa
    params::Option{T}              # system constants
end

Implicit representation of the nonlocal BEM system matrix.

Special constructors

Iterative solvers

NESSie.BEM._get_laplace_matrices — Function

_get_laplace_matrices{T}(
    Ξ       ::Vector{Vector{T}},
    elements::Vector{Triangle{T}}
)

Constructs and returns implicit representations for the single- and double-layer Laplace potential matrices V and K for the given observation points and surface elements.

Return type

Tuple{InteractionMatrix{T}, InteractionMatrix{T}}

NESSie.BEM._get_yukawa_matrices — Function

_get_yukawa_matrices{T}(
    Ξ       ::Vector{Vector{T}},
    elements::Vector{Triangle{T}},
    yuk     ::T
)

Constructs and returns implicit representations for the single- and double-layer Yukawa potential matrices V^Y and K^Y for the given observation points and surface elements.

Return type

Tuple{InteractionMatrix{T}, InteractionMatrix{T}}

NESSie.BEM._solve_linear_system — Function

_solve_linear_system{T}(
    A::AbstractArray{T, 2},
    b::AbstractArray{T, 1}
)

Solves the linear system Ax = b from A and b using a Jacobi-preconditioned GMRES solver.

Return type

Array{T, 1}