NESSie.BEM

NESSie.BEM.Kfun

NESSie.BEM.Kyfun

NESSie.BEM.LocalSystemMatrix

NESSie.BEM.NonlocalSystemMatrix

NESSie.BEM.Vfun

NESSie.BEM.Vyfun

NESSie.BEM._get_laplace_matrices

NESSie.BEM._get_yukawa_matrices

NESSie.BEM._solve_explicit

NESSie.BEM._solve_implicit

NESSie.BEM._solve_linear_system

Solvers

NESSie.BEM._get_laplace_matrices — Function

_get_laplace_matrices(
    Ξ       ::Vector{Vector{T}},
    elements::Vector{Triangle{T}}
)

Constructs and returns implicit representations for the single- and double-layer Laplace potential matrices V and K for the given observation points and surface elements.

Return type

Tuple{InteractionMatrix{T}, InteractionMatrix{T}}

NESSie.BEM._get_yukawa_matrices — Function

_get_yukawa_matrices(
    Ξ       ::Vector{Vector{T}},
    elements::Vector{Triangle{T}},
    yuk     ::T
)

Constructs and returns implicit representations for the single- and double-layer Yukawa potential matrices V^Y and K^Y for the given observation points and surface elements.

Return type

Tuple{InteractionMatrix{T}, InteractionMatrix{T}}

NESSie.BEM._solve_explicit — Function

_solve_explicit(
         ::Type{<: LocalityType},
    model::Model{T, Triangle{T}}
)

Computes the full local or nonlocal cauchy data on the surface of the biomolecule using an explicit representation of the BEM system.

Return type

LocalBEMResult{T, Triangle{T}} or NonlocalBEMResult{T, Triangle{T}}

NESSie.BEM._solve_implicit — Function

_solve_implicit(
         ::Type{<: LocalityType},
    model::Model{T, Triangle{T}}
)

Computes the full local or nonlocal cauchy data on the surface of the biomolecule using an implicit representation of the BEM system.

Return type

LocalBEMResult{T, Triangle{T}} or NonlocalBEMResult{T, Triangle{T}}

NESSie.BEM._solve_linear_system — Function

_solve_linear_system(
    A::AbstractArray{T, 2},
    b::AbstractArray{T, 1}
)

Solves the linear system Ax = b from A and b using a Jacobi-preconditioned GMRES solver.

Return type

Array{T, 1}

Supported keyword arguments

Everything from IterativeSolvers.gmres except for log.

Implicit potential matrices

NESSie.BEM.Kfun — Type

struct Kfun{T} <: InteractionFunction{Vector{T}, Triangle{T}, T}
    dat::Vector{T}   # pre-allocated vector for internal use
end

Interaction function for an implicit representation of potential matrix K.

Special constructors

Kfun{T}()

Automatically initializes the internal data vector. Replaces the default constructor.

NESSie.BEM.Kyfun — Type

struct Kyfun{T} <: InteractionFunction{Vector{T}, Triangle{T}, T}
    yuk::T   # exponent of the Yukawa operator's fundamental solution
end

Interaction function for an implicit representation of potential matrix Kʸ.

NESSie.BEM.Vfun — Type

struct Vfun{T} <: InteractionFunction{Vector{T}, Triangle{T}, T}
    dat::Vector{T}   # pre-allocated vector for internal use
end

Interaction function for an implicit representation of potential matrix V.

Special constructors

Vfun{T}()

Automatically initializes the internal data vector. Replaces the default constructor.

NESSie.BEM.Vyfun — Type

struct Vyfun{T} <: InteractionFunction{Vector{T}, Triangle{T}, T}
    yuk::T   # exponent of the Yukawa operator's fundamental solution
end

Interaction function for an implicit representation of potential matrix Vʸ.

Implicit system matrices

NESSie.BEM.LocalSystemMatrix — Type

struct LocalSystemMatrix{T} <: AbstractArray{T, 2}
    K     ::InteractionMatrix{T}   # double-layer Laplace
    params::Option{T}              # System constants
end

Implicit representation of the first local BEM system matrix.

NESSie.BEM.NonlocalSystemMatrix — Type

struct NonlocalSystemMatrix{T} <: AbstractArray{T, 2}
    V     ::InteractionMatrix{T}   # single-layer Laplace
    K     ::InteractionMatrix{T}   # double-layer Laplace
    Vy    ::InteractionMatrix{T}   # single-layer Yukawa
    Ky    ::InteractionMatrix{T}   # double-layer Yukawa
    params::Option{T}              # system constants
end

Implicit representation of the nonlocal BEM system matrix.

Special constructors

function NonlocalSystemMatrix{T}(
    Ξ       ::Vector{Vector{T}},
    elements::Vector{Triangle{T}},
    params  ::Option{T}
) where T

Creates the nonlocal system matrix directly from the given observation points Ξ and surface elements.