NESSie.Format

Input formats

NESSie.Format.readhmo_nodes — Function

readhmo_nodes(
    stream::IOStream,
          ::Type{T}=Float64
)

Reads all nodes from the given HMO file.

Return type

Vector{Vector{T}}

NESSie.Format.readhmo_elements — Function

readhmo_elements(
    stream::IOStream,
    nodes ::Vector{Vector{T}}
)

Reads all elements from the given HMO file.

Return type

Vector{Triangle{T}}

NESSie.Format.readhmo_charges — Function

readhmo_charges(
    stream::IOStream,
          ::Type{T}=Float64
)

Reads all charges from the given HMO file.

Return type

Vector{Charge{T}}

NESSie.Format.readmcsf_nodes — Function

readmcsf_nodes(
    stream::IOStream,
          ::Type{T}=Float64
)

Reads all nodes from the given GAMer-generated mcsf file.

Return type

Vector{Vector{T}}

NESSie.Format.readmcsf_elements — Function

readmcsf_elements(
    stream::IOStream,
    nodes ::Vector{Vector{T}};
    # kwargs
    domain::Symbol=:none
)

Reads all elements from the given GAMer-generated mcsf file.

Return type

Vector{Tetrahedron{T}}

NESSie.Format.readmsms_nodes — Function

readmsms_nodes(
    stream::IOStream,
          ::Type{T}=Float64
)

Reads all nodes from the given MSMS-generated .vert file.

Return type

Vector{Vector{T}}

NESSie.Format.readmsms_elements — Function

readmsms_elements(
    stream::IOStream,
    nodes ::Vector{Vector{T}}
)

Reads all elements from the given MSMS-generated .face file.

Return type

Vector{Triangle{T}}

NESSie.Format.readoff_nodes — Function

readoff_nodes(
    stream::IOStream,
    n     ::Int,
          ::Type{T}=Float64
)

Reads the first n nodes from the given OFF file.

Return type

Vector{Vector{T}}

NESSie.Format.readoff_elements — Function

readoff_elements(
    stream::IOStream,
    n     ::Int,
    nodes ::Vector{Vector{T}},
          ::Type{T}=Float64
)

Reads the first n elements from the given OFF file.

Return type

Vector{Triangle{T}}