NESSie.Radon

NESSie.Radon.radoncoll — Function

radoncoll(
    ξ       ::AbstractVector{T},
    tquad   ::TriangleQuad{T},
    yukawa  ::T,
    solution::Function
)

Seven-point Radon cubature [Rad48] for a given function and a pair of surface element and observations point.

If you intend computing single/double layer potentials with this function, you might want to use the shorthand signature regularyukawacoll instead.

Note

The result is premultiplied by 4π.

Arguments

solution Fundamental solution; supported functions: regularyukawapot, ∂ₙregularyukawapot

yukawa Exponent of the Yukawa operator's fundamental solution

Return type

Void

NESSie.Radon.radoncoll! — Function

radoncoll!(
        dest    ::AbstractVector{T},
        elements::AbstractVector{Triangle{T}},
        Ξ       ::AbstractVector{Vector{T}},
        solution::Function,
        yukawa  ::T,
        fvals   ::AbstractVector{T}
)

radoncoll!(
        dest    ::AbstractMatrix{T},
        elements::AbstractVector{Triangle{T}},
        Ξ       ::AbstractVector{Vector{T}},
        solution::Function,
        yukawa  ::T
)

Seven-point Radon cubature [Rad48] for a given function and a list of triangles and observation points Ξ. If dest is a vector, the function values f for each surface triangle have to be specified, since each element of the vector represents the dot product of the corresponding coefficient matrix row and the fvals vector.

If you intend computing single/double layer potentials with this function, you might want to use the shorthand signature regularyukawacoll! instead.

Note

The result is premultiplied by 4π.

Arguments

solution Fundamental solution; supported functions: regularyukawapot, ∂ₙregularyukawapot

yukawa Exponent of the Yukawa operator's fundamental solution

Return type

Void

NESSie.Radon.regularyukawapot — Function

regularyukawapot(
    x     ::AbstractVector{T},
    ξ     ::AbstractVector{T},
    yukawa::T,
          ::AbstractVector{T}=T[]
)

Computes the regular part of the Yukawa potential, that is, Yukawa minus Laplace:

\[\mathcal{G}^Y-\mathcal{G}^L = \frac{1}{4π}\frac{e^{-\frac{|x - ξ|}{Λ}} - 1}{|x-ξ|}\]

Note

The result is premultiplied by 4π.

Arguments

yukawa Exponent of the Yukawa operator's fundamental solution

Return type

T

NESSie.Radon.∂ₙregularyukawapot — Function

∂ₙregularyukawapot(
    x     ::AbstractVector{T},
    ξ     ::AbstractVector{T},
    yukawa::T,
    normal::AbstractVector{T}
)

Computes the normal derivative of the regular part of the Yukawa potential, that is, Yukawa minus Laplace:

\[\frac{∂}{∂n} \frac{1}{4π} \frac{e^{-\frac{|x - ξ|}{Λ}} - 1}{|x-ξ|} = \frac{1}{4π} \frac{1 - (1 + Λ^{-1} |x - ξ|)e^{-\frac{|x - ξ|}{Λ}}}{|x-ξ|²} \frac{(x - ξ) ⋅ n}{|x - ξ|}\]

Note

The result is premultiplied by 4π.

Arguments

yukawa Exponent of the Yukawa operator's fundamental solution

Return type

T