NESSie.TestModel

NESSie.TestModel._espotential_Σ

NESSie.TestModel._espotential_Ω

NESSie.TestModel._rfpotential_Σ

NESSie.TestModel._rfpotential_Ω

NESSie.TestModel._xie_coefficients

NESSie.TestModel.legendre

NESSie.TestModel.scalemodel

NESSie.TestModel.spherical_besseli

NESSie.TestModel.spherical_besselk

Potentials

NESSie.TestModel._espotential_Ω — Function

_espotential_Ω(::Type{<: LocalityType}, ξ::Vector{T}, ion::BornIon{T})
_espotential_Ω(::Type{<: LocalityType}, Ξ::AbstractVector{T}, ion::BornIon{T})

Computes the local or nonlocal electrostatic potential for (an) observation point(s) ξ (Ξ) inside the Born sphere.

Supported keyword arguments

See molpotential

Unit

$V = \frac{C}{F}$

Return type

T or Vector{T}

_espotential_Ω(ξ::Vector{T}, xie::XieTestModel{T})
_espotential_Ω(Ξ::AbstractVector{T}, xie::XieTestModel{T})

Computes the local or nonlocal electrostatic potential for (an) observation point(s) ξ (Ξ) inside the test model sphere.

Supported keyword arguments

See molpotential

Unit

$V = \frac{C}{F}$

Return type

T or Vector{T}

NESSie.TestModel._espotential_Σ — Function

_espotential_Σ(::Type{<: LocalityType}, ξ::Vector{T}, ion::BornIon{T})
_espotential_Σ(::Type{<: LocalityType}, Ξ::AbstractVector{T}, ion::BornIon{T})

Computes the local or nonlocal electrostatic potential for (an) observation point(s) ξ (Ξ) outside the Born sphere.

Supported keyword arguments

See molpotential

Unit

$V = \frac{C}{F}$

Return type

T or Vector{T}

_espotential_Σ(ξ::Vector{T}, xie::XieTestModel{T})
_espotential_Σ(Ξ::AbstractVector{T}, xie::XieTestModel{T})

Computes the local or nonlocal electrostatic potential for (an) observation point(s) ξ (Ξ) outside the test model sphere.

Supported keyword arguments

See molpotential

Unit

$V = \frac{C}{F}$

Return type

T or Vector{T}

NESSie.TestModel._rfpotential_Ω — Function

_rfpotential_Ω(::Type{<: LocalityType}, ξ::Vector{T}, ion::BornIon{T})
_rfpotential_Ω(::Type{<: LocalityType}, Ξ::AbstractVector{T}, ion::BornIon{T})

Computes the local or nonlocal reaction field potential for (an) observation point(s) ξ (Ξ) inside the Born sphere.

Unit

$V = \frac{C}{F}$

Return type

T or Vector{T}

_rfpotential_Ω(ξ::Vector{T}, xie::XieTestModel{T})
_rfpotential_Ω(Ξ::AbstractVector{T}, xie::XieTestModel{T})

Computes the local or nonlocal reaction field potential for (an) observation point(s) ξ (Ξ) inside the test model sphere.

Unit

$V = \frac{C}{F}$

Return type

T or Vector{T}

NESSie.TestModel._rfpotential_Σ — Function

_rfpotential_Σ(::Type{<: LocalityType}, ξ::Vector{T}, ion::BornIon{T})
_rfpotential_Σ(::Type{<: LocalityType}, Ξ::AbstractVector{T}, ion::BornIon{T})

Computes the local or nonlocal reaction field potential for (an) observation point(s) ξ (Ξ) outside the Born sphere.

Supported keyword arguments

See molpotential

Unit

$V = \frac{C}{F}$

Return type

T or Vector{T}

_rfpotential_Σ(ξ::Vector{T}, xie::XieTestModel{T})
_rfpotential_Σ(Ξ::AbstractVector{T}, xie::XieTestModel{T})

Computes the local or nonlocal reaction field potential for (an) observation point(s) ξ (Ξ) outside the test model sphere.

Supported keyword arguments

See molpotential

Unit

$V = \frac{C}{F}$

Return type

T or Vector{T}

Utility functions

NESSie.TestModel._xie_coefficients — Function

function _xie_coefficients(
    ::XieTestModel{T}
    model::XieSphere{T},
    len  ::Int
)

Depending on the concrete test model, computes the coefficients $A_{in}$ or $C_{in}$ with $i=1, 2, 3$ for the given XieSphere and the desired number of terms.

Return type

NTuple{3, Matrix{T}}

NESSie.TestModel.legendre — Function

legendre(
    maxn::Int,
    x   ::T
)

Precomputes the Legendre polynomials $Pₙ(x)$ for $n = 0, 1, ..., \texttt{maxn}-1$ and returns a generic function to access the values.

Return type

(generic function)

(n::Int) -> Pₙ(x)   # return type: T

Example

julia> p = legendre(3, 0.5);

julia> [p(n) for n in 0:2]    # [P₀(0.5), P₁(0.5), P₂(0.5)]
3-element Vector{Float64}:
  1.0
  0.5
 -0.125

NESSie.TestModel.scalemodel — Function

function scalemodel(
    charges::Vector{Charge{T}},
    radius ::T;
    # kwargs
    compat ::Bool              = false
)

Translates and rescales the given point charge model to fit inside an origin-centered sphere with the specified radius. More specifically, the model is centered at the origin and, if it contains more than one charge, scaled in a way such that the outermost point charge will be located at 80% radius distance from the origin.

Arguments

compat Enables compatibility mode and scales the model exactly like the reference implementation ([Xie16]). Use this flag if you intend to compare the results to the reference.

Return type

Vector{Charge{T}}

NESSie.TestModel.spherical_besseli — Function

spherical_besseli(
    maxn::Int,
    r   ::T
)

Precomputes the modified spherical Bessel function of the first kind $iₙ(r)$ for $n=-1, 0, ..., \texttt{maxn}$ and returns a generic function to access the values. $iₙ(r)$ is defined as

\[iₙ(r) = \sqrt{\frac{π}{2r}} I_{n+0.5}(r),\]

where $I_ν$ is the modified Bessel function of the first kind [Xie16].

Return type

(generic function)

(n::Int) -> iₙ(r)   # return type: T

NESSie.TestModel.spherical_besselk — Function

spherical_besselk(
    maxn::Int,
    r   ::T
)

Precomputes the modified spherical Bessel function of the second kind $kₙ(r)$ for $n=-1, 0, ..., \texttt{maxn}$ and returns a generic function to access the values. $kₙ(r)$ is defined as

\[kₙ(r) = \sqrt{\frac{π}{2r}} K_{n+0.5}(r),\]

where $K_ν$ is the modified Bessel function of the second kind [Xie16].

Return type

(generic function)

(n::Int) -> kₙ(r)   # return type: T