Utility functions

NESSie.cathetus — Function

cathetus{T}(hyp::T, cosθ::T)

Computes the cathetus $c₁$ of a triangle given the hypotenuse $h$ and the cosine of the exterior angle $θ$ between the hypotenuse and the other cathetus $c₂$.

\[c₂ = h ⋅ \cos(θ) \\ h² = c₁² + c₂² \\ ⇔ c₁ = \sqrt{h² ⋅ (1 - \cos²(θ))}\]

Return type

T

NESSie.TestModel.coefficients — Function

function coefficients{T}(
    model::XieModel{T},
    len  ::Int
)

Computes the coefficients $A_{in}$ with $i=1, 2, 3$ for the given XieModel and the desired number of terms.

Return type

Tuple{ Array{T, 2}, Array{T, 2}, Array{T, 2} }

Base.cos — Function

cos{T}(
    u    ::Vector{T},
    v    ::Vector{T},
    unorm::T=norm(u),
    vnorm::T=norm(v)
)

Computes the cosine of the angle between the given vectors u and v with lengths unorm and vnorm, respectively.

Return type

T

NESSie.ddot — Function

ddot{T}(
    u::Vector{T},
    v::Vector{T},
    n::Vector{T}
)

Devectorized computation of (u-v)⋅n.

Return type

T

NESSie.distance — Function

distance{T}(
    q   ::Vector{T},
    elem::Triangle{T}
)

Calculates the (positive or negative) distance from the given point q to the plane the given triangle elem is located in.

Return type

T

NESSie.eye! — Function

eye!{T}(
    m::Union{DenseArray{T,2}, SubArray{T,2}},
    α::Number=one(T)
)

Initializes the given matrix m with αI, with I being an identity matrix with the same dimensions as m.

Return type

Void

Example

julia> m = 2 * ones(2, 2)
2×2 Array{Float64,2}:
 2.0  2.0
 2.0  2.0

julia> eye!(m); m
2×2 Array{Float64,2}:
 1.0  0.0
 0.0  1.0

julia> eye!(m, 2); m
2×2 Array{Float64,2}:
 2.0  0.0
 0.0  2.0

NESSie.isdegenerate — Function

isdegenerate{T}(elem::Triangle{T})

Tests whether the given triangle is degenerate.

Return type

Bool

NESSie.TestModel.legendre — Function

legendre{T <: AbstractFloat}(
    maxn::Int,
    x   ::T
)

Precomputes the Legendre polynomials $Pₙ(x)$ for $n = 0, 1, ..., \texttt{maxn}-1$ and returns a generic function to access the values.

Return type

(generic function)

(n::Int) -> Pₙ(x)   # return type: T

Example

julia> p = legendre(3, 0.5);

julia> [p(n) for n in 0:2]    # [P₀(0.5), P₁(0.5), P₂(0.5)]
3-element Array{Float64,1}:
  1.0
  0.5
 -0.125

NESSie.pluseye! — Function

pluseye!{T}(
    m::Union{DenseArray{T,2}, SubArray{T,2}},
    α::Number=one(T)
)

Adds α to all diagonal elements of matrix m.

Return type

Void

Example

julia> m = 2 * ones(2, 2)
2×2 Array{Float64,2}:
 2.0  2.0
 2.0  2.0

julia> pluseye!(m); m
2×2 Array{Float64,2}:
 3.0  2.0
 2.0  3.0

julia> pluseye!(m, 2); m
2×2 Array{Float64,2}:
 5.0  2.0
 2.0  5.0

NESSie.props — Function

props{T}(
    elem::Triangle{T}
)

Computes the given triangle's properties, that is, centroid, normal, distance to origin, and area. Returns the completely initialized Triangle as a copy.

Warning

The given triangle remains unchanged!

Return type

Triangle

NESSie.reverseindex — Function

reverseindex{T}(v::Vector{T})

Creates a reverse index for the given vector v, that is, a dictionary linking the object IDs of the vector elements to the corresponding position in the vector.

Return type

Dict{UInt, UInt}

NESSie.TestModel.scalemodel — Function

function scalemodel{T}(
    charges::Vector{Charge{T}},
    radius ::T;
    # kwargs
    compat ::Bool              = false
)

Translates and rescales the given point charge model to fit inside an origin-centered sphere with the specified radius. More specifically, the function will center the given model at the origin and scaled in a way such that the outermost point charge will be located at 80% radius distance from the origin.

Arguments

compat Enables compatibility mode and scales the model exactly like the reference implementation ([Xie16]). Use this flag if you intend to compare the results to the reference.

Return type

Vector{Charge{T}}

Base.seek — Function

seek(
    fh         ::IOStream,
    prefix     ::String,
    skiptheline::Bool = true
)

Fast-forwards an IOStream to the next line starting with the given prefix. In case there is no such line, the stream handle will be set to EOF.

Arguments

skiptheline If true, said line will also be skipped

Return type

Void

Base.sign — Function

sign{T}(
    u::Vector{T},
    v::Vector{T},
    n::Vector{T}
)

Determines whether the normal vector of the plane specified by the vectors u and v has the same orientation as the given normal vector n. Returns $1$ if both normals have the same orientation, $0$ if at least one of the vectors is zero, and $-1$ otherwise.

Return type

T

NESSie.TestModel.spherical_besseli — Function

spherical_besseli{T <: AbstractFloat}(
    maxn::Int,
    r   ::T
)

Precomputes the modified spherical Bessel function of the first kind $iₙ(r)$ for $n=-1, 0, ..., \texttt{maxn}$ and returns a generic function to access the values. $iₙ(r)$ is defined as

\[iₙ(r) = \sqrt{\frac{π}{2r}} I_{n+0.5}(r),\]

where $I_ν$ is the modified Bessel function of the first kind [Xie16].

Return type

(generic function)

(n::Int) -> iₙ(r)   # return type: T

NESSie.TestModel.spherical_besselk — Function

spherical_besselk{T <: AbstractFloat}(
    maxn::Int,
    r   ::T
)

Precomputes the modified spherical Bessel function of the second kind $kₙ(r)$ for $n=-1, 0, ..., \texttt{maxn}$ and returns a generic function to access the values. $kₙ(r)$ is defined as

\[kₙ(r) = \sqrt{\frac{π}{2r}} K_{n+0.5}(r),\]

where $K_ν$ is the modified Bessel function of the second kind [Xie16].

Return type

(generic function)

(n::Int) -> kₙ(r)   # return type: T

NESSie.unpack — Function

unpack{T}(data::Vector{Vector{T}})

Unpacks the given vector of vectors into a single vector.

Return type

Vector{T}

Example

julia> unpack([[1, 2], [3]])
3-element Array{Int64,1}:
 1
 2
 3

NESSie.vertexnormals — Function

vertexnormals{T}(model::Model{T, Triangle{T}})

Returns a vector containing the normal vectors of the given model's triangles.

Return type

Vector{Vector{T}}