Electrostatics

NESSie.DoubleLayer

NESSie.LocalES

NESSie.LocalityType

NESSie.NonlocalES

NESSie.PotentialType

NESSie.SingleLayer

NESSie.espotential

NESSie.molpotential

NESSie.rfenergy

NESSie.rfpotential

Potential types

NESSie.PotentialType — Type

abstract type PotentialType end
struct SingleLayer <: PotentialType end
struct DoubleLayer <: PotentialType end

Enum-like representation of single and double layer potentials

NESSie.SingleLayer — Type

struct SingleLayer <: PotentialType end

Enum-like representation of single layer potentials

NESSie.DoubleLayer — Type

struct DoubleLayer <: PotentialType end

Enum-like representation of double layer potentials

Locality assumption

NESSie.LocalityType — Type

abstract type LocalityType end
struct NonlocalES <: LocalityType end
struct LocalES    <: LocalityType end

Enum-like representation of locality assumptions

NESSie.LocalES — Type

struct LocalES <: LocalityType end

Enum-like representation of local electrostatics problems, assuming complete independenc of solvent molecules

NESSie.NonlocalES — Type

struct NonlocalES <: LocalityType end

Enum-like representation of nonlocal electrostatics problems, allowing solvent molecule correlation effects (with area-of-effect radius λ)

Potentials

NESSie.espotential — Function

espotential(::Type{<: LocalityType}, ξ::Vector{T}, ion::BornIon{T})
espotential(::Type{<: LocalityType}, Ξ::AbstractArray{Vector{T}}, ion::BornIon{T})

Computes the local or nonlocal electrostatic potential(s) at the given observation point(s) ξ (Ξ). This function automatically locates the observation point(s).

The electrostatic potential is computed as the sum of the corresponding reaction field potential and the molecular potential.

Supported keyword arguments

See molpotential

Unit

$V = \frac{C}{F}$

Return type

T or Array{T}

Alias

espotential(domain::Symbol, ::Type{<: LocalityType}, ξ::Vector{T}, ::BornIon{T})espotential(domain::Symbol, ::Type{<: LocalityType}, Ξ::AbstractArray{Vector{T}}, ::BornIon{T})

Computes the electrostatic potential(s) for the given observation point(s) ξ (Ξ) and the given domain :Ω, :Σ, or :Γ.

espotential(ξ::Vector{T}, xie::XieTestModel{T})
espotential(Ξ::AbstractArray{Vector{T}}, xie::XieTestModel{T})

Computes the local or nonlocal electrostatic potential(s) at the given observation point(s) ξ (Ξ).

The electrostatic potential is computed as the sum of the corresponding reaction field potential and the molecular potential.

Supported keyword arguments

See molpotential

Unit

$V = \frac{C}{F}$

Return type

T or Array{T}

Alias

espotential(domain::Symbol, ξ::Vector{T}, xie::XieTestModel{T})espotential(domain::Symbol, Ξ::AbstractArray{Vector{T}}, xie::XieTestModel{T})

Computes the electrostatic potential(s) for the given observation point(s) ξ (Ξ) and the given domain :Ω, :Σ, or :Γ.

espotential(ξ::Vector{T}, bem::BEMResult{T})
espotential(Ξ::AbstractArray{Vector{T}}, bem::BEMResult{T})

Computes the local or nonlocal electrostatic potential(s) at the given observation point(s) ξ (Ξ). This function tries to automatically locate the observation point(s) using guess_domain.

The electrostatic potential is computed as the sum of the corresponding reaction field potential and the molecular potential.

Supported keyword arguments

surface_margin::T = 1e-6 see guess_domain

tolerance::T = 1e-10 see molpotential

Unit

$V = \frac{C}{F}$

Return type

T or Array{T}

Alias

espotential(domain::Symbol, ξ::Vector{T}, bem::BEMResult{T})espotential(domain::Symbol, Ξ::AbstractArray{Vector{T}}, bem::BEMResult{T})

Computes the electrostatic potential(s) for the given observation point(s) ξ (Ξ) and the given domain :Ω, :Σ, or :Γ.

NESSie.molpotential — Function

molpotential(ξ::Vector{T}, ion::BornIon{T})
molpotential(Ξ::AbstractArray{Vector{T}}, ion::BornIon{T})

Computes the molecular potential(s) at the given observation point(s) ξ (Ξ).

Supported keyword arguments

tolerance::T = 1e-10 minimum distance assumed between any observation point and point charge. Closer distances are replaced by this value.

Unit

$V = \frac{C}{F}$

Return type

T or Array{T}

molpotential(ξ::Vector{T}, xie::XieSphere{T})
molpotential(ξ::Vector{T}, xie::XieTestModel{T})
molpotential(Ξ::AbstractArray{Vector{T}}, xie::XieSphere{T})
molpotential(Ξ::AbstractArray{Vector{T}}, xie::XieTestModel{T})

Computes the molecular potential(s) at the given observation point(s) ξ (Ξ).

Supported keyword arguments

tolerance::T = 1e-10 minimum distance assumed between any observation point and point charge. Closer distances are replaced by this value.

Unit

$V = \frac{C}{F}$

Return type

T or Array{T}

molpotential(ξ::Vector{T}, bem::BEMResult{T})
molpotential(Ξ::AbstractArray{Vector{T}}, bem::BEMResult{T})

Computes the molecular potential(s) at the given observation point(s) ξ (Ξ).

Supported keyword arguments

tolerance::T = 1e-10 minimum distance assumed between any observation point and point charge. Closer distances are replaced by this value.

Unit

$V = \frac{C}{F}$

Return type

T or Array{T}

molpotential(ξ::Vector{T}, model::Model{T})
molpotential(Ξ::AbstractArray{Vector{T}}, model::Model{T})

Computes the molecular potential(s) at the given observation point(s) ξ (Ξ) for the given model.

Supported keyword arguments

tolerance::T = 1e-10 minimum distance assumed between any observation point and point charge. Closer distances are replaced by this value.

Unit

$V = \frac{C}{F}$

Return type

T or Array{T}

NESSie.rfpotential — Function

rfpotential(::Type{<: LocalityType}, ξ::Vector{T}, ion::BornIon{T})
rfpotential(::Type{<: LocalityType}, Ξ::AbstractArray{Vector{T}}, ion::BornIon{T})

Computes the local or nonlocal reaction field potential(s) at the given observation point(s) ξ (Ξ). This function automatically locates the observation point(s).

Supported keyword arguments

See molpotential

Unit

$V = \frac{C}{F}$

Return type

T or Array{T}

Alias

rfpotential(domain::Symbol, ::Type{<: LocalityType}, ξ::Vector{T}, ::BornIon{T})rfpotential(domain::Symbol, ::Type{<: LocalityType}, Ξ::AbstractArray{Vector{T}}, ::BornIon{T})

Computes the reaction field potential(s) for the given observation point(s) ξ (Ξ) and the given domain :Ω, :Σ, or :Γ.

rfpotential(ξ::Vector{T}, xie::XieTestModel{T})
rfpotential(Ξ::AbstractArray{Vector{T}}, xie::XieTestModel{T})

Computes the local or nonlocal reaction field potential(s) at the given observation point(s) ξ (Ξ).

Supported keyword arguments

See molpotential

Unit

$V = \frac{C}{F}$

Return type

T or Array{T}

Alias

rfpotential(domain::Symbol, ξ::Vector{T}, xie::XieTestModel{T})rfpotential(domain::Symbol, Ξ::AbstractArray{Vector{T}}, xie::XieTestModel{T})

Computes the reaction field potential(s) for the given observation point(s) ξ (Ξ) and the given domain :Ω, :Σ, or :Γ.

rfpotential(ξ::Vector{T}, bem::BEMResult{T})
rfpotential(Ξ::AbstractArray{Vector{T}}, bem::BEMResult{T})

Computes the local or nonlocal reaction field potential(s) at the given observation point(s) ξ (Ξ). This function tries to automatically locate the observation point(s) using guess_domain.

Supported keyword arguments

surface_margin::T = 1e-6 see guess_domain

tolerance::T = 1e-10 see molpotential

Unit

$V = \frac{C}{F}$

Return type

T or Array{T}

Alias

rfpotential(domain::Symbol, ξ::Vector{T}, bem::BEMResult{T})rfpotential(domain::Symbol, Ξ::AbstractArray{Vector{T}}, bem::BEMResult{T})

Computes the reaction field potential(s) for the given observation point(s) ξ (Ξ) and the given domain :Ω, :Σ, or :Γ.

Potential energies

NESSie.rfenergy — Function

rfenergy(::Type{<: LocalityType}, ::BornIon{T})

Computes the local or nonlocal reaction field energy W*

\[W^* = ∫φ^* ρ \quad dΩ\]

where $φ^*$ is the reaction field and $ρ$ is the corresponding charge distribution.

Unit

$\frac{kJ}{mol}$

Return type

T

rfenergy(::XieTestModel{T})

Computes the local or nonlocal reaction field energy W*

\[W^* = ∫φ^* ρ \quad dΩ\]

where $φ^*$ is the reaction field and $ρ$ is the corresponding charge distribution.

Unit

$\frac{kJ}{mol}$

Return type

T

rfenergy(::BEMResult{T})

Computes the local or nonlocal reaction field energy W*

\[W^* = ∫φ^* ρ \quad dΩ\]

where $φ^*$ is the reaction field and $ρ$ is the corresponding charge distribution.

Unit

$\frac{kJ}{mol}$

Return type

T

abstract type LocalityType end struct NonlocalES <: LocalityType end struct LocalES <: LocalityType end
Enum-like representation of locality assumptions

mutable struct Model{T, E <: Element{T}} nodes ::Vector{Vector{T} = Vector{T}[] # mesh nodes elements::Vector{E} = E[] # mesh elements charges ::Vector{Charge{T}} = Charge{T}[] # point charges in the molecule params ::Option{T} = defaultopt(T) # system constants end
System model representing a biomelecule in solvation, including a collection of point charges in the molecule and a set of system constants.

mutable struct BornIon{T <: AbstractFloat} charge::Charge{T} # point charge at the sphere center radius::T # sphere radius in Å params::Option{T} = defaultopt(BornIon{T}) # system constants end
Single Born ion, that is, a monoatomic ion represented as a spherically symmetric domain with a single point charge located at its center ().

mutable struct XieSphere{T} radius ::T # radius of the origin-centered sphere charges::Vector{Charge{T}} # point charges in the sphere params ::Option{T} = defaultopt(T) # system constants end
System model of a dielectric sphere containing multiple point charges.

struct XieTestModel{T, M} end
Common abstraction for all Xie sphere-based test models.

abstract type BEMResult{T, E <: SurfaceElement{T}} end
Abstract base type for all BEM solver results