Electrostatics

NESSie.DoubleLayer

NESSie.LocalES

NESSie.LocalityType

NESSie.NonlocalES

NESSie.PotentialType

NESSie.SingleLayer

NESSie.espotential

NESSie.molpotential

NESSie.rfenergy

NESSie.rfpotential

Potential types

NESSie.PotentialType — Type

abstract type PotentialType end
struct SingleLayer <: PotentialType end
struct DoubleLayer <: PotentialType end

Enum-like representation of single and double layer potentials

NESSie.SingleLayer — Type

abstract type PotentialType end
struct SingleLayer <: PotentialType end
struct DoubleLayer <: PotentialType end

Enum-like representation of single and double layer potentials

NESSie.DoubleLayer — Type

abstract type PotentialType end
struct SingleLayer <: PotentialType end
struct DoubleLayer <: PotentialType end

Enum-like representation of single and double layer potentials

Locality assumption

NESSie.LocalityType — Type

abstract type LocalityType end
struct NonlocalES <: LocalityType end
struct LocalES    <: LocalityType end

Enum-like representation of locality assumption:

Local electrostatics: Complete independence of solvent molecules

Nonlocal electrostatics: Allow solvent molecule correlation effects (with area-of-effect radius λ)

NESSie.LocalES — Type

abstract type LocalityType end
struct NonlocalES <: LocalityType end
struct LocalES    <: LocalityType end

Enum-like representation of locality assumption:

Local electrostatics: Complete independence of solvent molecules

Nonlocal electrostatics: Allow solvent molecule correlation effects (with area-of-effect radius λ)

NESSie.NonlocalES — Type

abstract type LocalityType end
struct NonlocalES <: LocalityType end
struct LocalES    <: LocalityType end

Enum-like representation of locality assumption:

Local electrostatics: Complete independence of solvent molecules

Nonlocal electrostatics: Allow solvent molecule correlation effects (with area-of-effect radius λ)

Potentials

NESSie.espotential — Function

espotential(::Type{<: LocalityType}, ξ::Vector{T}, ion::BornIon{T})
espotential(::Type{<: LocalityType}, Ξ::AbstractVector{Vector{T}}, ion::BornIon{T})

Computes the local or nonlocal electrostatic potential(s) at the given observation point(s) ξ (Ξ) for the given born ion. This function automatically locates the observation point(s).

The electrostatic potential is computed as the sum of the corresponding reaction field potential and the molecular potential.

Supported keyword arguments

See molpotential

Unit

$V = \frac{C}{F}$

Return type

T or Vector{T}

Alias

espotential(domain::Symbol, ::Type{<: LocalityType}, ξ::Vector{T}, ::BornIon{T})
espotential(domain::Symbol, ::Type{<: LocalityType}, Ξ::AbstractVector{Vector{T}}, ::BornIon{T})

Computes the electrostatic potential(s) for the given observation point(s) ξ (Ξ) and the given domain :Ω, :Σ, or :Γ.

espotential(ξ::Vector{T}, xie::XieTestModel{T})
espotential(Ξ::AbstractVector{Vector{T}}, xie::XieTestModel{T})

Computes the local or nonlocal electrostatic potential(s) at the given observation point(s) ξ (Ξ) for the given Xie test model.

The electrostatic potential is computed as the sum of the corresponding reaction field potential and the molecular potential.

Supported keyword arguments

See molpotential

Unit

$V = \frac{C}{F}$

Return type

T or Vector{T}

Alias

espotential(domain::Symbol, ξ::Vector{T}, xie::XieTestModel{T})
espotential(domain::Symbol, Ξ::AbstractVector{Vector{T}}, xie::XieTestModel{T})

Computes the electrostatic potential(s) for the given observation point(s) ξ (Ξ) and the given domain :Ω, :Σ, or :Γ.

espotential(ξ::Vector{T}, bem::BEMResult{T})
espotential(Ξ::AbstractVector{Vector{T}}, bem::BEMResult{T})

Computes the local or nonlocal electrostatic potential(s) at the given observation point(s) ξ (Ξ) for the given BEM result. This function tries to automatically locate the observation point(s) using guess_domain.

The electrostatic potential is computed as the sum of the corresponding reaction field potential and the molecular potential.

Supported keyword arguments

surface_margin::T = 1e-6 see guess_domain

tolerance::T = 1e-10 see molpotential

Unit

$V = \frac{C}{F}$

Return type

T or Vector{T}

Alias

espotential(domain::Symbol, ξ::Vector{T}, bem::BEMResult{T})
espotential(domain::Symbol, Ξ::AbstractVector{Vector{T}}, bem::BEMResult{T})

Computes the electrostatic potential(s) for the given observation point(s) ξ (Ξ) and the given domain :Ω, :Σ, or :Γ.

NESSie.molpotential — Function

molpotential(ξ::Vector{T}, ion::BornIon{T})
molpotential(Ξ::AbstractVector{Vector{T}}, ion::BornIon{T})

Computes the molecular potential(s) at the given observation point(s) ξ (Ξ) for the given born ion.

Supported keyword arguments

tolerance::T = 1e-10 minimum distance assumed between any observation point and point charge. Closer distances are replaced by this value.

Unit

$V = \frac{C}{F}$

Return type

T or Vector{T}

molpotential(ξ::Vector{T}, xie::XieSphere{T})
molpotential(ξ::Vector{T}, xie::XieTestModel{T})
molpotential(Ξ::AbstractVector{Vector{T}}, xie::XieSphere{T})
molpotential(Ξ::AbstractVector{Vector{T}}, xie::XieTestModel{T})

Computes the molecular potential(s) at the given observation point(s) ξ (Ξ) for the given Xie sphere or Xie test model.

Supported keyword arguments

tolerance::T = 1e-10 minimum distance assumed between any observation point and point charge. Closer distances are replaced by this value.

Unit

$V = \frac{C}{F}$

Return type

T or Vector{T}

molpotential(ξ::Vector{T}, model::Model{T})
molpotential(Ξ::AbstractVector{Vector{T}}, model::Model{T})

Computes the molecular potential(s) at the given observation point(s) ξ (Ξ) for the given model.

Supported keyword arguments

tolerance::T = 1e-10 minimum distance assumed between any observation point and point charge. Closer distances are replaced by this value.

Unit

$V = \frac{C}{F}$

Return type

T or Vector{T}

molpotential(ξ::Vector{T}, bem::BEMResult{T})
molpotential(Ξ::AbstractVector{Vector{T}}, bem::BEMResult{T})

Computes the molecular potential(s) at the given observation point(s) ξ (Ξ) for the given BEM result.

Supported keyword arguments

tolerance::T = 1e-10 minimum distance assumed between any observation point and point charge. Closer distances are replaced by this value.

Unit

$V = \frac{C}{F}$

Return type

T or Vector{T}

NESSie.rfpotential — Function

rfpotential(::Type{<: LocalityType}, ξ::Vector{T}, ion::BornIon{T})
rfpotential(::Type{<: LocalityType}, Ξ::AbstractVector{Vector{T}}, ion::BornIon{T})

Computes the local or nonlocal reaction field potential(s) at the given observation point(s) ξ (Ξ) for the given born ion. This function automatically locates the observation point(s).

Supported keyword arguments

See molpotential

Unit

$V = \frac{C}{F}$

Return type

T or Vector{T}

Alias

rfpotential(domain::Symbol, ::Type{<: LocalityType}, ξ::Vector{T}, ::BornIon{T})
rfpotential(domain::Symbol, ::Type{<: LocalityType}, Ξ::AbstractVector{Vector{T}}, ::BornIon{T})

Computes the reaction field potential(s) for the given observation point(s) ξ (Ξ) and the given domain :Ω, :Σ, or :Γ.

rfpotential(ξ::Vector{T}, xie::XieTestModel{T})
rfpotential(Ξ::AbstractVector{Vector{T}}, xie::XieTestModel{T})

Computes the local or nonlocal reaction field potential(s) at the given observation point(s) ξ (Ξ) for the given Xie test model.

Supported keyword arguments

See molpotential

Unit

$V = \frac{C}{F}$

Return type

T or Vector{T}

Alias

rfpotential(domain::Symbol, ξ::Vector{T}, xie::XieTestModel{T})
rfpotential(domain::Symbol, Ξ::AbstractVector{Vector{T}}, xie::XieTestModel{T})

Computes the reaction field potential(s) for the given observation point(s) ξ (Ξ) and the given domain :Ω, :Σ, or :Γ.

rfpotential(ξ::Vector{T}, bem::BEMResult{T})
rfpotential(Ξ::AbstractVector{Vector{T}}, bem::BEMResult{T})

Computes the local or nonlocal reaction field potential(s) at the given observation point(s) ξ (Ξ) for the given BEM result. This function tries to automatically locate the observation point(s) using guess_domain.

Supported keyword arguments

surface_margin::T = 1e-6 see guess_domain

tolerance::T = 1e-10 see molpotential

Unit

$V = \frac{C}{F}$

Return type

T or Vector{T}

Alias

rfpotential(domain::Symbol, ξ::Vector{T}, bem::BEMResult{T})
rfpotential(domain::Symbol, Ξ::AbstractVector{Vector{T}}, bem::BEMResult{T})

Computes the reaction field potential(s) for the given observation point(s) ξ (Ξ) and the given domain :Ω, :Σ, or :Γ.

Potential energies

NESSie.rfenergy — Function

rfenergy(::Type{<: LocalityType}, ::BornIon{T})

Computes the local or nonlocal reaction field energy W* as

\[W^* = ∫φ^* ρ \quad dΩ\]

where $φ^*$ is the reaction field and $ρ$ is the corresponding charge distribution.

Unit

$\frac{kJ}{mol}$

Return type

T

rfenergy(::XieTestModel{T})

Computes the local or nonlocal reaction field energy W* as

\[W^* = ∫φ^* ρ \quad dΩ\]

where $φ^*$ is the reaction field and $ρ$ is the corresponding charge distribution.

Unit

$\frac{kJ}{mol}$

Return type

T

rfenergy(::BEMResult{T})

Computes the local or nonlocal reaction field energy W* as

\[W^* = ∫φ^* ρ \quad dΩ\]

where $φ^*$ is the reaction field and $ρ$ is the corresponding charge distribution.

Unit

$\frac{kJ}{mol}$

Return type

T