Solvers

BEM solvers

NESSie.BEM.BEMResult — Type

abstract type BEMResult{T, E <: SurfaceElement{T}} end

Abstract base type for all BEM solver results

NESSie.BEM.LocalBEMResult — Type

struct LocalBEMResult{T, E} <: BEMResult{T, E}
    model::Model{T, E}
    u    ::Vector{T}   # [γ₀int(φ*)](ξ)    ∀ ξ ∈ Ξ; premultiplied by 4π⋅ε0
    q    ::Vector{T}   # [γ₁int(φ*)](ξ)    ∀ ξ ∈ Ξ; premultiplied by 4π⋅ε0
    umol ::Vector{T}   # [γ₀int(φ*mol)](ξ) ∀ ξ ∈ Ξ; premultiplied by 4π⋅ε0
    qmol ::Vector{T}   # [γ₁int(φ*mol)](ξ) ∀ ξ ∈ Ξ; premultiplied by 4π⋅ε0
end

Result data of the local solving process to be used for potential computation and post-processing, with Ξ being the list of observation points, that is, the set of triangle centroids.

NESSie.BEM.NonlocalBEMResult — Type

struct NonlocalBEMResult{T, E} <: BEMResult{T, E}
    model::Model{T, E}
    u    ::SubArray{T,1}   # [γ₀int(φ*)](ξ)    ∀ ξ ∈ Ξ; premultiplied by 4π⋅ε0
    q    ::SubArray{T,1}   # [γ₁int(φ*)](ξ)    ∀ ξ ∈ Ξ; premultiplied by 4π⋅ε0
    w    ::SubArray{T,1}   # [γ₀ext(Ψ)](ξ)     ∀ ξ ∈ Ξ; premultiplied by 4π⋅ε0
    umol ::Vector{T}       # [γ₀int(φ*mol)](ξ) ∀ ξ ∈ Ξ; premultiplied by 4π⋅ε0
    qmol ::Vector{T}       # [γ₁int(φ*mol)](ξ) ∀ ξ ∈ Ξ; premultiplied by 4π⋅ε0
end

Result data of the nonlocal solving process to be used for potential computation and post-processing, with Ξ being the list of observation points, that is, the set of triangle centroids.

NESSie.BEM.solve — Function

solve{T, L <: LocalityType}(
         ::L,
    model::Model{T, Triangle{T}}
)

Computes the full local or nonlocal cauchy data on the surface of the biomolecule using an explicit representation of the BEM system.

Note

The explicit system representations require a substantial amount of memory and might easily exceed the available resources on your machine (the memory demands scale quadratically with the number of surface elements). If you encounter this problem, please consider switching to solve_implicit or our CUDA-accelerated solvers from the CuNESSie.jl package instead, which provide the same function interface and are freely interchangeable.

Return type

LocalBEMResult{T, Triangle{T}} or NonlocalBEMResult{T, Triangle{T}}

NESSie.BEM.solve_implicit — Function

solve_implicit{T, L <: LocalityType}(
         ::L,
    model::Model{T, Triangle{T}}
)

Computes the full local or nonlocal cauchy data on the surface of the biomolecule using an implicit representation of the BEM system.

Note

The implicit system representations provide a small memory footprint even for large biomolecular systems (the memory demands scale only linearly with the number of surface elements). However, this comes at the cost of additional computation time, which might exceed several hours, depending on the system size. If you encounter this problem, please consider switching to solve or our CUDA-accelerated solvers from the CuNESSie.jl package instead, which provide the same function interface and are freely interchangeable.

Return type

LocalBEMResult{T, Triangle{T}} or NonlocalBEMResult{T, Triangle{T}}