Utility functions

GeometryBasics.Rect

GeometryBasics.mesh

NESSie.guess_domain

NESSie.meshunion

NESSie.nessie_data_path

NESSie.obspoints_plane

GeometryBasics.Rect — Type

GeometryBasics.Rect(model::Model{T, Triangle{T}})

Creates and returns a bounding box for the given model.

Supported keyword arguments

padding::T = 0 Adds padding to each side of the (otherwise flush) bounding box, effectively increasing its size by 2 ⋅ padding in each direction.

Return type

GeometryBasics.Rect3{T}

GeometryBasics.mesh — Function

GeometryBasics.mesh(model::Model{T, Triangle{T}})

Converts the given model into a GeometryBasics.jl-compatible mesh, e.g., for visualization through Makie.jl.

Return type

GeometryBasics.Mesh

NESSie.guess_domain — Function

guess_domain(ξ::Vector{T}, model::Model{T, Triangle{T}})

Make an educated guess whether the given observation point ξ is located in the protein domain Ω, the solvent domain Σ, or the surface domain Γ.

The domain is determined based on the relative position to the triangle with the closest centroid.

Supported keyword arguments

surface_margin::T = 1e-3 maximum |cos| ⋅ norm allowed between the vector from the closest triangle's centroid to ξ and the triangle's unit normal vector before ξ is no longer considered to be part of the Γ domain.

Return type

Symbol (:Ω, :Σ, or :Γ)

Alias

guess_domain(Ξ::AbstractArray{Vector{T}}, model::Model{T, Triangle{T}})

Determines the domain of each observation point ξ ∈ Ξ.

NESSie.meshunion — Function

meshunion(
    model1::Model{T, Tetrahedron{T}},
    model2::Model{T, Tetrahedron{T}}
)

Merges two volume models, e.g., the models of a protein and the solvent. Duplicate nodes (e.g., the nodes on the protein surface) are merged into a single node, duplicate elements and charges (if any) are retained as well as the system constants of the first model.

Return type

Model{T, Tetrahedron{T}}

Note

This function assumes that there are no duplicates within either of the node lists!

NESSie.nessie_data_path — Function

nessie_data_path(parts::AbstractString...)

Returns the absolute path to NESSie's "data" directory, if used without arguments, or to the location relative to that, specified by parts (e.g., a directory within "data").

Return type

String

NESSie.obspoints_plane — Function

obspoints_plane(
    a  ::AbstractVector{T},
    b  ::AbstractVector{T},
    c  ::AbstractVector{T},
    nba::Int,
    nbc::Int
)

Generates nba ⋅ nbc evenly distributed observation points on the parallelogram with the sides $\overline{BA}$ and $\overline{BC}$ and a, b, c being the location vectors to the points $A$, $B$, and $C$, respectively.

Arguments

nba Number of observation points along $\overline{BA}$

nbc Number of observation points along $\overline{BC}$

Return type

Generator -> LinRange{Vector{T}, Int}

Example

for Ξ in obspoints_plane(...)    for ξ in Ξ        ...    endend

mutable struct Model{T, E <: Element{T}} nodes ::Vector{Vector{T} = Vector{T}[] # mesh nodes elements::Vector{E} = E[] # mesh elements charges ::Vector{Charge{T}} = Charge{T}[] # point charges in the molecule params ::Option{T} = defaultopt(T) # system constants end
System model representing a biomelecule in solvation, including a collection of point charges in the molecule and a set of system constants.

struct Tetrahedron{T} <: VolumeElement{T} v1::Vector{T} # position of the first node v2::Vector{T} # position of the second node v3::Vector{T} # position of the third node v4::Vector{T} # position of the fourth node domain::Symbol # element domain (solvent :Σ, solute :Ω, or :none) end
Representation of a single tetrahedron.

struct Triangle{T} <: SurfaceElement{T} v1 ::Vector{T} # position of the first node v2 ::Vector{T} # position of the second node v3 ::Vector{T} # position of the third node center ::Vector{T} # centroid of the triangle normal ::Vector{T} # normal vector of the triangle area ::T # area of the triangle distorig::T # distance to the origin end
Representation of a single surface triangle.